requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.7
langchain-google-genai==2.0.4
langchain-experimental==0.3.3
pandas==2.2.3
matplotlib==3.9.2
streamlit==1.39.0
python-dotenv==1.0.1
tabulate==0.9.0
